pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "browser-harness"
version = "0.1.0"
description = "The simplest, thinnest, and most powerful harness to control your real browser with your agent."
requires-python = ">=3.11"
dependencies = [
    "cdp-use==1.4.5",
    "fastapi>=0.136.1",
    "fetch-use==0.4.0",
    "google-genai>=1.73.1",
    "pydantic>=2.13.3",
    "uvicorn>=0.46.0",
    "websockets==16.0",
]

[project.scripts]
browser-harness = "run:main"

[tool.setuptools]
py-modules = ["run", "helpers", "daemon", "admin"]
